remove code duplication
... and introduce separate kernel extension for pre-loading JDBC drivers
